#322: refactoring to split up write.py into individual files
